feat(icons): consolidated function
